Add a note about complianceType
This adds a link to the complianceType documentation and spells out a
difference that was entirely non-obvious to me: 'musthave' does not
enforce the full object, it just makes sure the fields of the objects
are all there. I.e. it does a merge with an existing object not a
replace.